+++ desktop/packages/mullvad-vpn/src/renderer/components/ProblemReport.tsx
@@ -26,7 +26,7 @@ import { useEffectEvent } from '../lib/utility-hooks';
 import { useSelector } from '../redux/store';
 import support from '../redux/support/actions';
 import { AppNavigationHeader } from './';
-import { BackAction } from './KeyboardNavigation';
+import { BackAction } from './keyboard-navigation';
 import { Footer, Layout, SettingsContainer } from './Layout';
 import { ModalAlert, ModalAlertType } from './Modal';
 import {