+++ desktop/packages/mullvad-vpn/src/renderer/components/views/changelog/ChangelogView.tsx
@@ -4,7 +4,7 @@ import { useHistory } from '../../../lib/history';
 import { useVersionCurrent } from '../../../redux/hooks';
 import { AppNavigationHeader } from '../../';
 import { ChangelogList } from '../../changelog-list';
-import { BackAction } from '../../KeyboardNavigation';
+import { BackAction } from '../../keyboard-navigation';
 import { Layout, SettingsContainer } from '../../Layout';
 import { NavigationContainer } from '../../NavigationContainer';
 import { NavigationScrollbars } from '../../NavigationScrollbars';