+++ desktop/packages/mullvad-vpn/src/renderer/hooks/useIsPlatform.ts
@@ -1,0 +1,7 @@
+type Platform = Extract<NodeJS.Platform, 'linux' | 'darwin' | 'win32'>;
+
+export const useIsPlatform = (platform: Platform) => {
+  const isPlatform = window.env.platform === platform;
+
+  return isPlatform;
+};