+++ CHANGELOG.md
@@ -34,6 +34,13 @@ Line wrap the file at 100 chars.                                              Th
 #### Linux
 - Upgrade split-tunneling to use cgroups v2, instead of the deprecated cgroups v1.
   Users on Linux kernels prior to 5.13 will not be able to use split tunneling.
+- Change "Go back" keyboard shortcut to `Ctrl + ←` or `Ctrl + [`.
+
+#### macOS
+- Change "Go back" keyboard shortcut to `Cmd + ←` or `Cmd + [`.
+
+#### Windows
+- Change "Go back" keyboard shortcut to `Ctrl + ←` or `Ctrl + [`.
 
 
 ## [2025.14] - 2025-12-09